api/requirements.txt
pymongo
pydantic

apifairy==1.3.0
    # via -r requirements.in
faker==19.12.0
    # via -r requirements.in
flask==3.0.0
    # via
    #   -r requirements.in
    #   apifairy
    #   flask-cors
    #   flask-httpauth
    #   flask-mail
    #   flask-marshmallow
    #   flask-sqlalchemy
flask-cors==4.0.0
    # via -r requirements.in
flask-httpauth==4.8.0
    # via
    #   -r requirements.in
    #   apifairy
flask-mail==0.9.1
    # via -r requirements.in
flask-marshmallow==0.15.0
    # via
    #   -r requirements.in
    #   apifairy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.20.1
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-sqlalchemy==0.29.0
    # via -r requirements.in
packaging==23.2
    # via
    #   apispec
    #   build
    #   flask-marshmallow
    #   gunicorn
    #   marshmallow
    #   marshmallow-sqlalchemy
    #   webargs
pip-tools==7.3.0
    # via -r requirements.in
pyjwt==2.8.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
werkzeug==3.0.1
    # via flask
